fix: Enhance blog functionality with author details, public view fetching, and improved error messages

+++ src/app/api/(blogs)/blog/[id]/route.ts
@@ -3,19 +3,13 @@ import dbConnect from "@/lib/dbConnect";
 import Blog from "@/model/Blog";
 import { NextRequest, NextResponse } from "next/server";
 
+const permittedDesignations = ["Director", "Assistant Director"];
+const permittedDepartments = ["Press Release and Publications"];
+
 export async function GET(
   request: NextRequest,
   { params }: { params: { id: string } },
 ) {
-  const { session, isPermitted } = await hasAuth();
-
-  if (!session || !isPermitted) {
-    return NextResponse.json(
-      { message: "You are not authorized to view this page" },
-      { status: 401 },
-    );
-  }
-
   try {
     await dbConnect();
     const { id } = params;
@@ -35,11 +29,14 @@ export async function PATCH(
   request: NextRequest,
   { params }: { params: { id: string } },
 ) {
-  const { session, isPermitted } = await hasAuth();
+  const { session, isPermitted } = await hasAuth(
+    permittedDesignations,
+    permittedDepartments,
+  );
 
-  if (!session || !isPermitted) {
+  if (!session) {
     return NextResponse.json(
-      { message: "You are not authorized to view this page" },
+      { message: "You are not authorized to update this blog" },
       { status: 401 },
     );
   }
@@ -49,16 +46,28 @@ export async function PATCH(
     const { id } = params;
     const updateData = await request.json();
 
-    const blog = await Blog.findByIdAndUpdate(id, updateData, {
-      new: true,
-      runValidators: true,
-    });
-
+    // Check if the user is allowed to update this blog
+    const blog = await Blog.findById(id);
     if (!blog) {
       return NextResponse.json({ error: "Blog not found" }, { status: 404 });
     }
 
-    return NextResponse.json(blog, { status: 200 });
+    // Allow updates only if the user is the author or has permissions
+    const isAuthor =
+      blog.author.authorId.toString() === session.user.id.toString();
+    if (!isAuthor && !isPermitted) {
+      return NextResponse.json(
+        { message: "You are not authorized to update this blog" },
+        { status: 403 },
+      );
+    }
+
+    const updatedBlog = await Blog.findByIdAndUpdate(id, updateData, {
+      new: true,
+      runValidators: true,
+    });
+
+    return NextResponse.json(updatedBlog, { status: 200 });
   } catch (error: any) {
     return NextResponse.json({ error: error.message }, { status: 500 });
   }
@@ -68,11 +77,14 @@ export async function DELETE(
   request: NextRequest,
   { params }: { params: { id: string } },
 ) {
-  const { session, isPermitted } = await hasAuth();
+  const { session, isPermitted } = await hasAuth(
+    permittedDesignations,
+    permittedDepartments,
+  );
 
-  if (!session || !isPermitted) {
+  if (!session) {
     return NextResponse.json(
-      { message: "You are not authorized to view this page" },
+      { message: "You are not authorized to delete this blog" },
       { status: 401 },
     );
   }
@@ -81,10 +93,23 @@ export async function DELETE(
     await dbConnect();
     const { id } = params;
 
-    const blog = await Blog.findByIdAndDelete(id);
+    const blog = await Blog.findById(id);
     if (!blog) {
       return NextResponse.json({ error: "Blog not found" }, { status: 404 });
     }
+
+    // Allow deletions only if the user is the author or has permissions
+    const isAuthor =
+      blog.author.authorId.toString() === session.user.id.toString();
+
+    if (!isAuthor && !isPermitted) {
+      return NextResponse.json(
+        { message: "You are not authorized to delete this blog" },
+        { status: 403 },
+      );
+    }
+
+    await Blog.findByIdAndDelete(id);
 
     return NextResponse.json(
       { message: "Blog deleted successfully" },
